update profile state management

+++ user/delete_user_wishlist.php
@@ -1,15 +1,25 @@
 <?php
 session_start();
 require_once('../controllers/dbconn.php');
+require_once('../includes/current_user_data.php');
 
 $wishlist_id = $_GET['id'];
 
 if (isset($wishlist_id)) {
 
-    $query = "delete from wishlist where wishlist_id = ?;";
+
+    $stmt = $conn->prepare("SELECT * FROM wishlist WHERE wishlist_id = ? AND user_id = ?");
+    $stmt->execute([$wishlist_id, $user_id]);
+
+    if ($stmt->rowCount() === 0) {
+        $_SESSION['error'] = 'Unauthorized access';
+        header("Location: ../user/profile.php");
+        exit;
+    }
+
+    $query = "delete from wishlist where wishlist_id = ?";
     $stmt = $conn->prepare($query);
     $status =  $stmt->execute([$wishlist_id]);
-
 
 
     if ($status) {
